Set weight in matchmaking types to an optimal default value if all types should be played equally

+++ botli_dataclasses.py
@@ -198,8 +198,8 @@ class Matchmaking_Type:
     rated: bool
     variant: Variant
     perf_type: Perf_Type
-    multiplier: int
-    weight: int
+    multiplier: float
+    weight: float
     min_rating_diff: int
     max_rating_diff: int
